+++ clang/docs/ReleaseNotes.rst
@@ -1113,6 +1113,7 @@ OpenMP Support
 - Added support 'no_openmp_constructs' assumption clause.
 - Added support for 'self_maps' in map and requirement clause.
 - Added support for 'omp stripe' directive.
+- Added support for threadset clause in task and taskloop directives.
 - Fixed a crashing bug with ``omp unroll partial`` if the argument to
   ``partial`` was an invalid expression. (#GH139267)
 - Fixed a crashing bug with ``omp tile sizes`` if the argument to ``sizes`` was